schema: regenerate for the 2026-09-23 update
server.json is a fresh Source2SchemaDumper 1.0.20 dump (dump_schema all
for_cpp) taken on the updated Linux server. Removed classes (CCSPlayerCamera,
CEnvWindController, CEnvWindVolume, CLogicPlayerProxy, CBaseMoveBehavior, ...)
and fields (CPhysicsProp::m_bThrownByPlayer, ...) are gone from the headers,
so a stale use fails to compile instead of resolving to offset 0.

generator.py still emitted the META_RES form of AddSingleEntityIOListener;
it now matches the Action signature the hand edit introduced.

+++ public/source2toolkit/schema/entity/enums/VMixFilterType_t.h
@@ -42,7 +42,7 @@
 
 #include <cstdint>
 
-enum class VMixFilterType_t : int16_t
+enum class VMixFilterType_t : int8_t
 {
     FILTER_UNKNOWN = -1,
     FILTER_LOWPASS = 0,
